Functioned it (at last)
Created functions (dot_plot, swarm_plot) that replace the many lines of code for creating swarm and dot plots.

For now, they cannot be fully generalized but it is a good start.

They can also be incorporated into the same function or in a general plotting class as they are quite similar.