Gradient Boosting integration
Added a new class in the models_class.py file. The integration has not been smooth and many problems have arose as a result.

This leads us to the redesign of the code structure in a separate branch for this project to address problems such as expandability, integration and ease of use.