Even more messing around with Pytorch
Tried to understand the basics and I think I got them.
I built a simple ANN based on the tutorials I have found to understand the commands and how to bring the network to life and it worked. We see a pretty low MSE but there are more things to be done here, if I want to integrate this script to the main pipeline