Integration of single input file and saving procedure to Gradient boosting
Integrated the single file input and the saving procedure to the Gradient Boosting algorithm.

The above includes (GradientBoosting_main):
- The introduction of some new parameters in certain methods,
- The creation of a variable responsible for the operating point (op)
- The change of the name of the file

The "main" script is controlled by the execution which handles data introduction. The parameters need to be altered, based on the hyperparameters chosen for the model. That is because, as of now, no hyperparameters are chosen or considered for this particular model.

The result of the execution has been left as an example.