Syntax problem - Check issue https://github.com/pantoniad/Correlate/issues/12
Problem identified with the execution file of the GBR, where the fed data to the learning curve method did not change based on the operating point. The problem was a small syntax error in the data given where the "_idle" part was not changed in the "df_final_<operating point>" variable that was passed. This resulted in the same learning curve being plotted again and again but with different parameter settings, based on the corresponding operating point.

The same problem has not been seen in the execution file of the polynomial regressor.

Follow issue suggestions.

Also kept the outputs for the execution of all three models.