Now saving the metrics of the complexity plots
For the generation of Taylor plots, data is required from a series of tests on the selected models. As we generate data for the creation of the creation plot, the same models generated will be used to showcase the improvement of the model on a Taylor Diagram.

For this, for the GBR and ANN:
- All needed types of metrics are calculated for the results of the complexity plots,
- The metrics are saved on a separate folder inside the "Results" folder,
- Additionally, the predictions for the specific engine tested, are now being saved on a separate folder.

Plots and metrics are kept for reference.